user routes with bug
